pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptguard-sdk"
version = "1.5.0"
description = "Drop-in security for AI applications - AI Firewall SDK"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "PromptGuard", email = "support@promptguard.co" }
]
keywords = [
    "ai",
    "security",
    "llm",
    "prompt-injection",
    "openai",
    "anthropic",
    "ai-firewall",
    "pii-detection",
    "llm-security",
    "ai-security",
    "langchain",
    "crewai",
    "llamaindex",
    "auto-instrumentation",
    "genai-security",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "httpx>=0.24.0",
]

[project.optional-dependencies]
langchain = [
    "langchain-core>=0.1.0",
]
crewai = [
    "crewai>=0.28.0",
]
llamaindex = [
    "llama-index-core>=0.10.0",
]
all = [
    "langchain-core>=0.1.0",
    "crewai>=0.28.0",
    "llama-index-core>=0.10.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://promptguard.co"
Documentation = "https://docs.promptguard.co"
Repository = "https://github.com/acebot712/promptguard-python"
Issues = "https://github.com/acebot712/promptguard-python/issues"

[tool.hatch.build.targets.wheel]
packages = ["promptguard"]

[tool.ruff]
line-length = 100
target-version = "py310"
fix = true

[tool.ruff.lint]
select = [
    "F",    # pyflakes
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # bugbear
    "C4",   # comprehensions
    "RUF",  # ruff-specific
]
ignore = []

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["promptguard"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true
